update the list view more intelligently
The list view is not updated after every command anymore, but only when the
buffer is reverted or if in discovery mode.  Properties of paired devices are
updated by the signal handler, otherwise ‘revert-buffer’ must be called to
update.  A mode line update is now forced in the adapter signal handler.

* bluetooth.el (bluetooth--update-timer): update timer
(bluetooth--update-all): new function, used in tabulated-list-revert-hook
(bluetooth--print-list): renamed from bluetooth--update-list
(bluetooth--update-print): new function